backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
functions-framework==3.4.0

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# AI and LangChain
langchain==0.0.350
langchain-anthropic==0.0.1
anthropic==0.7.7

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0 